requirements.txt
awsebcli==3.15.0
botocore==1.12.129
cement==2.8.2
certifi==2019.3.9
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.1.1
django-excel==0.0.10
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-simplejwt==4.2.0
docutils==0.14
et-xmlfile==1.0.1
future==0.16.0
idna==2.7
jdcal==1.4
jmespath==0.9.4
lml==0.0.9
Markdown==3.1
odfpy==1.4.0
openpyxl==2.5.14
pathspec==0.5.9
pyexcel==0.5.13
pyexcel-io==0.5.17
pyexcel-ods==0.5.6
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.7
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==1.6.1
urllib3==1.24.1
wcwidth==0.1.7
xlrd==1.2.0
xlwt==1.3.0
